:truck: (counter) move old counter files to history
