+++ Commands/History/CounterCategory.cs
@@ -1,0 +1,16 @@
+﻿namespace Bishop.Commands.History;
+
+/// <summary>
+///     Represent every possible category for <see cref="RecordEntity" />.
+///     <remarks>ORDERS OF THE KEYS MUST NOT CHANGE AS ENUMS ARE CONVERTED TO INT WITH KEY POSITION</remarks>
+/// </summary>
+public enum CounterCategory
+{
+    Bdm,
+    Sauce,
+    Sel,
+    Beauf,
+    Rass,
+    Malfoy,
+    Wind
+}
