+++ Commands/History/RecordRepository.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Bishop.Commands.Meter;
 using Bishop.Helper.Database;
 using MongoDB.Driver;
 using MongoDB.Driver.Linq;
