--- Commands/Meter/CounterCategory.cs
@@ -1,16 +1,0 @@
-﻿namespace Bishop.Commands.Meter;
-
-/// <summary>
-///     Represent every possible category for <see cref="CounterEntity" />.
-///     <remarks>ORDERS OF THE KEYS MUST NOT CHANGE AS ENUMS ARE CONVERTED TO INT WITH KEY POSITION</remarks>
-/// </summary>
-public enum CounterCategory
-{
-    Bdm,
-    Sauce,
-    Sel,
-    Beauf,
-    Rass,
-    Malfoy,
-    Wind
-}
