+++ Config/Converters/MeterKeysConverter.cs
@@ -1,5 +1,5 @@
 ﻿using System.Threading.Tasks;
-using Bishop.Commands.Meter;
+using Bishop.Commands.History;
 using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Converters;
 using DSharpPlus.Entities;
