+++ Program.cs
@@ -5,8 +5,7 @@ using System.Threading.Tasks;
 using Bishop.Commands.CardGame;
 using Bishop.Commands.Dump;
 using Bishop.Commands.History;
-using Bishop.Commands.Meter;
-using Bishop.Commands.Meter.Aliases;
+using Bishop.Commands.History.Aliases;
 using Bishop.Commands.Weather;
 using Bishop.Config;
 using Bishop.Helper;
